Fixed misc issues. Created WordCard component to reduce code duplication.
