+++ apps/client/src/app/dictionary-page/word-card-mobile.tsx
@@ -4,6 +4,7 @@ import {
   FullScreenModal,
 } from '@fv-app/common-components';
 import React from 'react';
+import WordCard from './word-card';
 
 function WordCardMobile({
   word,
@@ -14,9 +15,6 @@ function WordCardMobile({
   theme,
 }: FvWord) {
   const [showModal, setShowModal] = React.useState(false);
-
-  const secondaryButtonStyle = useButtonStyle('secondary', 'button');
-  const tertiaryButtonStyle = useButtonStyle('tertiary', 'button');
 
   return (
     <>
@@ -33,7 +31,9 @@ function WordCardMobile({
           </div>
           <div className="self-center col-span-2">
             {audio != null &&
-              audio.map((_element) => <i className="fv-volume-up" />)}
+              audio.map((fvAudio) => (
+                <i key={fvAudio.filename} className="fv-volume-up" />
+              ))}
           </div>
           <div className="place-self-end self-center">
             <i className="fv-right-open" />
@@ -64,41 +64,25 @@ function WordCardMobile({
         >
           <div className="p-10">
             <p className="grow font-bold text-3xl">{word}</p>
-            <p className="italic">
-              {optional != null &&
-                optional.map((element) => {
-                  if (element['Part of Speech'] !== null) {
-                    return `(${element['Part of Speech']}) `;
-                  } else {
-                    return '';
-                  }
-                })}
-            </p>
-            <p className="pt-10 pb-10">{definition}</p>
-            {audio != null &&
-              audio.map((fvAudio) => (
-                <button
-                  className={secondaryButtonStyle}
-                  onClick={() => playAudio(fvAudio.filename)}
-                >
-                  <i className="fv-play">{fvAudio.speaker}</i>
-                </button>
-              ))}
-            {img && (
-              <img className="pt-10 max-w-md max-h-md" src={img} alt={word} />
-            )}
-            <p className="pt-10">CATEGORIES</p>
-            <button className={tertiaryButtonStyle}>{theme}</button>
+            <WordCard
+              word={word}
+              definition={definition}
+              img={img}
+              optional={optional}
+              audio={audio}
+              theme={theme}
+              source={''}
+              entryID={''}
+              secondary_theme={null}
+              compare_form={''}
+              sort_form={''}
+              sorting_form={[]}
+            />
           </div>
         </FullScreenModal>
       )}
     </>
   );
-}
-
-async function playAudio(fileName: string) {
-  const audio = new Audio(fileName);
-  audio.play();
 }
 
 export default WordCardMobile;
